requirements.txt
fastapi
supabase
python-dotenv
worldometer